Domicilio Persona OneToOne, Modulo Domicilio Arreglado

+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/common/CommonController.java
@@ -41,7 +41,7 @@ public class CommonController<E, S extends CommonIService<E>> {
 
 	@GetMapping("/{id}")
 	@Transactional
-	public ResponseEntity<?> getOne(@PathVariable int id) {
+	public ResponseEntity<?> getOne(@PathVariable long id) {
 
 		try {
 
@@ -75,7 +75,7 @@ public class CommonController<E, S extends CommonIService<E>> {
 
 	@PutMapping("/{id}")
 	@Transactional
-	public ResponseEntity<?> put(@PathVariable int id, @RequestBody E personaForm) {
+	public ResponseEntity<?> put(@PathVariable long id, @RequestBody E personaForm) {
 
 		try {
 
@@ -91,7 +91,7 @@ public class CommonController<E, S extends CommonIService<E>> {
 	
 	@DeleteMapping("/{id}")
 	@Transactional
-	public ResponseEntity<?> delete(@PathVariable int id) {
+	public ResponseEntity<?> delete(@PathVariable long id) {
 
 		try {
